Move StringBuilder to its own file
